+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+## [3.9.0] 2022-03-08
+
 - New task with source tracked sandbox:
   - Do not allow to select a major org for dev or config
   - Open SF org if selected from the already connected list